+++ Libplanet.Tests/Blockchain/BlockChainTest.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections.Generic;
 using System.Collections.Immutable;
 using System.Linq;
+using System.Security.Cryptography;
 using System.Threading.Tasks;
 using Bencodex.Types;
 using Libplanet.Action;
@@ -65,6 +66,7 @@ namespace Libplanet.Tests.Blockchain
             _emptyTransaction = new List<Transaction<DumbAction>>();
             _validNext = Block<DumbAction>.Mine(
                     1,
+                    HashAlgorithmType.Of<SHA256>(),
                     1024,
                     _fx.GenesisBlock.TotalDifficulty,
                     _fx.GenesisBlock.Miner.Value,